requirements.txt
fastapi
uvicorn
requests
pillow
pytesseract
pdfplumber
pydantic
